Use event driven Linux watcher
