+++ gradle/dependencies.gradle
@@ -76,7 +76,7 @@ libraries.kryo =                [coordinates: 'com.esotericsoftware.kryo:kryo', 
 libraries.maven3 =              [coordinates: 'org.apache.maven:maven-core', version: '3.0.4', license: "Apache 2.0"]
 libraries.maven3_wagon_file =   [coordinates: 'org.apache.maven.wagon:wagon-file', version: '3.0.0', because: '3.1.0 of wagon-http seems to break Digest authentication', license: "Apache 2.0"]
 libraries.maven3_wagon_http =   [coordinates: 'org.apache.maven.wagon:wagon-http', version: libraries.maven3_wagon_file.version, license: "Apache 2.0"]
-libraries.nativePlatform =      [coordinates: 'net.rubygrapefruit:native-platform', version: '0.22-snapshot-20200225084142+0000', license: "Apache 2.0"]
+libraries.nativePlatform =      [coordinates: 'net.rubygrapefruit:native-platform', version: '0.22-snapshot-20200303091216+0000', license: "Apache 2.0"]
 libraries.nekohtml =            [coordinates: 'net.sourceforge.nekohtml:nekohtml', version: '1.9.22', license: "Apache 2.0"]
 libraries.objenesis =           [coordinates: 'org.objenesis:objenesis', version: '2.6', license: "Apache 2.0"]
 libraries.plexus_container =    [coordinates: 'org.codehaus.plexus:plexus-container-default', version: '1.7.1', license: "Apache 2.0"]
